requirements.txt
youtube-search-python
pytube
ytmusicapi
python-dotenv
tqdm